move testSelectIntoUpdateCount into RegressionTest class
